Created cards package and moved appropiate classes
