+++ src/main/java/no/uib/inf112/core/ui/actors/cards/CardActor.java
@@ -3,7 +3,7 @@ package no.uib.inf112.core.ui.actors.cards;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageTextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
-import no.uib.inf112.core.player.Card;
+import no.uib.inf112.core.map.cards.Card;
 import no.uib.inf112.core.ui.UIHandler;
 import no.uib.inf112.desktop.Main;
 import org.jetbrains.annotations.Nullable;
